% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{.getmode}
\alias{.getmode}
\title{.getmode}
\usage{
.getmode(v, ties_method = "random")
}
\arguments{
\item{v}{input}

\item{ties_method}{(character) Method for breaking ties. Must be "random" or "mean"}
}
\value{
A summarization with ties broken
}
\description{
.getmode
}
