% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffered_bbox.R
\name{get_buffered_bbox}
\alias{get_buffered_bbox}
\title{Buffered bounding box}
\usage{
get_buffered_bbox(locs)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}
}
\value{
A buffered bounding box (bbox object)
}
\description{
Get a buffered bounding box for set of locations
}
\examples{
\dontrun{
data(locs)

get_buffered_bbox(locs)
}
}
