% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{get_climate}
\alias{get_climate}
\title{Climate data}
\usage{
get_climate(locs, id.label, res = 0.5, path = tempdir(), method = "fast")
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{id.label}{(character) Column name of location ID}

\item{res}{(numeric) Required argument for geodata::worldclim_global(). Valid resolutions are 10, 5, 2.5, and 0.5 (minutes of a degree).}

\item{path}{(character) Path to location of data to extract}

\item{method}{(character) Method to extract data using terra::extract. Valid methods are "fast" and "precise"}
}
\value{
A data frame with summarized minimum temperature, maximum temperature, and total precipitation for each polygon in locs
}
\description{
Climate data
}
\examples{
\dontrun{
data(locs)

get_climate(locs, id.label = 'grid.id')
}
}
