% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Elevation}
\usage{
get_elevation(
  locs,
  path,
  id.label,
  break.every = 1000,
  buffer = 100,
  method = "fast"
)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{path}{(character) Path to location of data to extract}

\item{id.label}{(character) Column name of location ID}

\item{method}{(character) Method to extract data using terra::extract. Valid methods are "fast" and "precise"}
}
\description{
Process US elevation data. Before running this function, you must download the USA elevation data from Download TIF from http://www.cec.org/north-american-environmental-atlas/elevation-2023/. Unzip and call the path to the elevation_tif folder in the get_elevation() function.
}
\examples{
\dontrun{
data(locs)

dat <- get_elevation(locs, path = 'data/', id.label = 'grid.id')
}
}
