% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanFootprint.R
\name{get_footprint}
\alias{get_footprint}
\title{Human footprint}
\usage{
get_footprint(locs, id.label, year = 2009, path = tempdir(), method = "fast")
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{id.label}{(character) Column name of location ID}

\item{year}{(numeric) Year for geodata::footprint(). Valid years are 1993 or 2009.}

\item{path}{(character) Path to location of data to extract}

\item{method}{(character) Method to extract data using terra::extract(). Valid methods are "fast" and "precise"}
}
\value{
A data frame with summarized human footprint for each polygon in locs
}
\description{
Download and process human footprint data. Source: geodata::footprint()
}
\examples{
\dontrun{
data(locs)

get_footprint(locs, id.label = 'grid.id')
}
}
