% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_population}
\alias{get_population}
\title{Population density}
\usage{
get_population(
  locs,
  id.label,
  year = 2010,
  res = 0.5,
  path = tempdir(),
  method = "fast"
)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{id.label}{(character) Column name of location ID}

\item{year}{(numeric) Year for geodata::population(). Valid years are 2000, 2005, 2010, 2015, and 2020}

\item{res}{(numeric) Resolution for geodata::population(). Valid resolutions are 10, 5, 2.5, and 0.5 (minutes of a degree)}

\item{path}{(character) Path to location of data to extract}

\item{method}{(character) Method to extract data using terra::extract. Valid methods are "fast" and "precise"}
}
\value{
A data frame with summarized population density for each polygon in locs
}
\description{
Download and process population density data. Source: geodata::population()
}
\examples{
\dontrun{
data(locs)

dat <- get_population(locs, id.label = 'grid.id')
}
}
