% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roads.R
\name{get_roads}
\alias{get_roads}
\title{Roads}
\usage{
get_roads(locs, path, id.label)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{path}{(character) Path to location of data to extract}

\item{id.label}{(character) Column name of location ID}
}
\value{
A data frame with summarized roads for each polygon in locs
}
\description{
Process US road data. Before running this function, you must download from \href{https://geodata.bts.gov/datasets/0b6c2fd2e3ac40a7929cdff1d4cf604a_0/explore}{Transportation.gov}. Unzip, rename the folder to \emph{roads.gdb}, and call the path to the folder in the \code{get_roads()} function.
}
\examples{
\dontrun{
data(locs)

get_roads(locs, path = 'data/', id.label = 'grid.id')
}
}
