% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{get_soil}
\alias{get_soil}
\title{Soil}
\usage{
get_soil(
  locs,
  id.label,
  var = c("bdod", "cfvo", "clay", "nitrogen", "ocd", "phh2o", "sand", "silt", "soc"),
  depth = 5,
  stat = "mean",
  name = "",
  path = tempdir(),
  method = "fast"
)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{id.label}{(character) Column name of location ID}

\item{var}{(character) Required argument for \pkg{geodata}::\code{soil_world()}. Valid options are: "bdod", "cfvo", "clay", "nitrogen", "ocd", "ocs", "phh2o", "sand", "silt", "soc", "wrb".}

\item{depth}{(numeric) Required argument for \pkg{geodata}::\code{soil_world()}. Valid options are: 5, 15, 30, 60, 100, 200}

\item{stat}{(character) Required argument for \pkg{geodata}::\code{soil_world()}. Valid options are: "mean", "uncertainty", "Q0.05", "Q0.5", "Q0.95". Ignored if var="wrb"}

\item{name}{(character) Required argument for \pkg{geodata}::\code{soil_world()}. One of "Acrisols", "Albeluvisols", "Alisols", "Andosols", "Arenosols", "Calcisols", "Cambisols", "Chernozems", "Cryosols", "Durisols", "Ferralsols", "Fluvisols", "Gleysols", "Gypsisols", "Histosols", "Kastanozems", "Leptosols", "Lixisols", "Luvisols", "Nitisols", "Phaeozems", "Planosols", "Plinthosols", "Podzols", "Regosols", "Solonchaks", "Solonetz", "Stagnosols", "Umbrisols", "Vertisols". Only used when var="wrb"}

\item{path}{(character) Path to location of data to extract}

\item{method}{(character) Method to extract data using \pkg{terra}::\code{extract()}. Valid methods are "fast" and "precise"}
}
\value{
A data frame with summarized soil composition for each polygon in locs
}
\description{
Download and process soil data using \pkg{geodata}::\code{soil_world()}
}
\examples{
\dontrun{
data(locs)

get_soil(locs, id.label = 'grid.id')
}
}
