% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travelTime.R
\name{get_traveltime}
\alias{get_traveltime}
\title{Travel time}
\usage{
get_traveltime(
  locs,
  id.label,
  to = "city",
  size = 4,
  up = T,
  path = tempdir(),
  method = "fast"
)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{id.label}{(character) Column name of location ID}

\item{to}{(character) Required argument for geodata::travel_time(). Valid options are "city" or "port"}

\item{size}{(numeric) Required argument for geodata::travel_time(). A positive integer indicating the size of the city or port. Can be between 1 and 9 if to="city" or between 1 and 5 if to="port".}

\item{up}{(logical) Required argument for geodata::travel_time(). If TRUE the travel time to a city of the size chosen or larger is returned.}

\item{path}{(character) Path to location of data to extract}

\item{method}{(character) Method to extract data using terra::extract(). Valid methods are "fast" and "precise"}
}
\value{
A data frame with summarized travel time for each polygon in locs
}
\description{
Download and process travel time data. Source: geodata::travel_time()
}
\examples{
\dontrun{
data(locs)

get_traveltime(locs, id.label = 'grid.id')
}
}
