% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlines.R
\name{get_flowlines}
\alias{get_flowlines}
\title{Flowlines}
\usage{
get_flowlines(locs, path, id.label = "id")
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{path}{(character) Path to location of data to extract}

\item{id.label}{(character) Column name of location ID}
}
\value{
A data frame with summarized flowlines for each polygon in locs
}
\description{
Process US flowline data. Before running this function, you must download NHDPlusV21_NationalData_Seamless_Geodatabase_Lower48_07.7z from https://www.epa.gov/waterdata/nhdplus-national-data. Unzip and call the path to the NHDPlusNationalData folder in the get_flowlines() function.
}
\examples{
\dontrun{
data(locs)

dat <- get_flowlines(locs, path = 'data/', id.label = 'grid.id')
}
}
