% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roads.R
\name{get_roads}
\alias{get_roads}
\title{Roads}
\usage{
get_roads(locs, path, id.label)
}
\arguments{
\item{locs}{(sf) Polygons for which to summarize covariates (should be grid cells, watersheds, or buffered points)}

\item{path}{(character) Path to location of data to extract}

\item{id.label}{(character) Column name of location ID}
}
\value{
A data frame with summarized roads for each polygon in locs
}
\description{
Process US road data. Before running this function, you must download from https://geodata.bts.gov/datasets/0b6c2fd2e3ac40a7929cdff1d4cf604a_0/explore?location=23.778950%2C70.504834%2C3.54. Unzip, rename the folder roads.gdb, and call the path to the folder in the get_roads() function.
}
\examples{
\dontrun{
data(locs)

dat <- get_roads(locs, path = 'data/', id.label = 'grid.id')
}
}
