% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locs}
\alias{locs}
\title{First five letters of the alphabet}
\format{
An sf data frame with 2872 rows and 2 variables:
\describe{
  \item{grid.id}{integer from 1 to number of grid cells (2872)}
  \item{geometry}{sf geometry, POLYGONS}
  ...
}
}
\usage{
locs
}
\description{
A dataset containing first five letters of English alphabet
}
\keyword{datasets}
